% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitauthors.doc.R
\docType{package}
\name{knitauthors}
\alias{knitauthors}
\alias{knitauthors-package}
\title{Format authors and affiliations for knitr markdown files}
\description{
Format authors and affiliations for knitr markdown files
}
\author{
James Ware \email{j.ware@imperial.ac.uk}
}

