% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/formatAuthors.R
\name{formatAuthors}
\alias{formatAuthors}
\title{format authors}
\usage{
formatAuthors(authorTable, firstSymbol = "\\\\*", lastSymbol = "†",
  corrSymbol = "\\\\§",
  coFirstText = "These authors contributed equally to this work",
  coLastText = "These authors contributed equally to this work",
  corrAuthorText = "Corresponding author")
}
\arguments{
\item{authorTable}{the input data}
}
\value{
Returns markdown-formatted text
}
\description{
format authors
}
\examples{
authorTable <- read.delim("data-raw/authors.tab", sep="\\t",)
knitauthors::formatAuthors(authorTable)
}

